+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - No unreleased changes so far
+
+## [0.46.0] - 2022-08-02
 ### Added
 - Attributes for diesel cars
 - New status departureTimersStatus & chargingProfilesStatus
+- Support for python 3.10
+
+### Fixed
+- Build without image support now works without asking for requirements
 
 ## [0.45.1] - 2022-07-25
 ### Fixed
@@ -605,7 +611,8 @@ Minor fix in observer interface
 ## [0.1.0] - 2021-05-26
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.45.1...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-python/compare/v0.46.0...HEAD
+[0.46.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.46.0
 [0.45.1]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.45.1
 [0.45.0]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.45.0
 [0.44.2]: https://github.com/tillsteinbach/WeConnect-python/releases/tag/v0.44.2